Add constants

+++ core/src/main/scala/de/leanovate/jbj/ast/expr/value/IntegerConstExpr.scala
@@ -4,7 +4,7 @@ import de.leanovate.jbj.ast.{FilePosition, Expr}
 import de.leanovate.jbj.runtime.Context
 import de.leanovate.jbj.runtime.value.IntegerVal
 
-case class IntegerConstExpr(position: FilePosition,value: Int) extends Expr {
+case class IntegerConstExpr(position: FilePosition, value: Long) extends Expr {
   lazy val _value = IntegerVal(value)
 
   def eval(ctx: Context) = _value
