+++ core/src/main/scala/de/leanovate/jbj/runtime/value/StringVal.scala
@@ -35,7 +35,7 @@ case class StringVal(value: String) extends Value {
   def decr = this
 
   def getAt(index: Value) = index.unref match {
-    case IntegerVal(idx) => StringVal(value(idx).toString)
+    case IntegerVal(idx) => StringVal(value(idx.toInt).toString)
     case NumericVal(idx) => StringVal(value(idx.toInt).toString)
     case _ => UndefinedVal
   }
